[FIX][CORE] Moved FollowMode activity creation to core
It seems that the FollowingActivitiesManager was skipped during
migration. Indeed it contains questionable logic that should not be part
of the core. However this patch now moves at least the activity
generation part into the core.
