+++ de.fu_berlin.inf.dpp/src/de/fu_berlin/inf/dpp/project/internal/FollowingActivitiesManager.java
@@ -5,7 +5,6 @@ import de.fu_berlin.inf.dpp.activities.StopFollowingActivity;
 import de.fu_berlin.inf.dpp.annotations.Component;
 import de.fu_berlin.inf.dpp.awareness.AwarenessInformationCollector;
 import de.fu_berlin.inf.dpp.editor.FollowModeManager;
-import de.fu_berlin.inf.dpp.editor.IFollowModeListener;
 import de.fu_berlin.inf.dpp.session.AbstractActivityConsumer;
 import de.fu_berlin.inf.dpp.session.AbstractActivityProducer;
 import de.fu_berlin.inf.dpp.session.IActivityConsumer;
@@ -18,9 +17,12 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import org.apache.log4j.Logger;
 import org.picocontainer.Startable;
 
+// TODO evaluate what logic can moved to the core. The AwarenessInformationCollector is a singleton
+// Application Runtime component, only used for the Eclipse UI. It is not the best choice to move it
+// also to the core.
 /**
- * This manager is responsible for distributing knowledge about changes in follow modes between
- * session participants. It both produces and consumes activities.
+ * This manager is responsible for collection changes in follow modes between session participants.
+ * It consumes activities.
  *
  * @author Alexander Waldmann (contact@net-corps.de)
  */
@@ -35,22 +37,6 @@ public class FollowingActivitiesManager extends AbstractActivityProducer impleme
   private final ISarosSession session;
 
   private final AwarenessInformationCollector collector;
-
-  private final FollowModeManager followModeManager;
-
-  private final IFollowModeListener followModeListener =
-      new IFollowModeListener() {
-
-        @Override
-        public void stoppedFollowing(Reason reason) {
-          fireActivity(new StopFollowingActivity(session.getLocalUser()));
-        }
-
-        @Override
-        public void startedFollowing(User target) {
-          fireActivity(new StartFollowingActivity(session.getLocalUser(), target));
-        }
-      };
 
   private final IActivityConsumer consumer =
       new AbstractActivityConsumer() {
@@ -92,7 +78,6 @@ public class FollowingActivitiesManager extends AbstractActivityProducer impleme
       final FollowModeManager followModeManager) {
     this.session = session;
     this.collector = collector;
-    this.followModeManager = followModeManager;
   }
 
   @Override
@@ -101,7 +86,6 @@ public class FollowingActivitiesManager extends AbstractActivityProducer impleme
     session.addActivityProducer(this);
     session.addActivityConsumer(consumer, Priority.ACTIVE);
     session.addListener(sessionListener);
-    followModeManager.addListener(followModeListener);
   }
 
   @Override
@@ -109,7 +93,6 @@ public class FollowingActivitiesManager extends AbstractActivityProducer impleme
     session.removeActivityProducer(this);
     session.removeActivityConsumer(consumer);
     session.removeListener(sessionListener);
-    followModeManager.removeListener(followModeListener);
     collector.flushFollowModes();
   }
